modify the code from review
