+++ src/components/todo/Item.js
@@ -1,6 +1,6 @@
 import { useRef, useState } from "react";
 import styled from "styled-components";
-import Todo from "./Todo";
+
 // Hint: props로 id, text, onClick 등 (자유)
 export default function Item({ title, index, deleteItem }) {
   const [isChecked, setIsChecked] = useState(false);
